FIX: ./philo 4 310 200 100 3 멈춤 해결

+++ philosophers/philo_bonus/simulation.c
@@ -6,7 +6,7 @@
 /*   By: mher <mher@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/19 16:00:29 by mher              #+#    #+#             */
-/*   Updated: 2022/06/27 16:38:52 by mher             ###   ########.fr       */
+/*   Updated: 2022/06/28 17:26:03 by mher             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,22 +18,29 @@ static int	return_with_free(pid_t *pids, int return_value)
 	return (return_value);
 }
 
-int	wait_end(t_philo *philo, pid_t *pids, pthread_t *full_monitor)
+int	wait_end(t_philo *philo, pid_t *pids, pthread_t full_monitor)
 {
-	pid_t			pid;
+	pid_t			first_dead_philo;
 	int				status;
 	unsigned int	i;
 
-	pid = waitpid(-1, &status, 0);
+	first_dead_philo = waitpid(-1, &status, 0);
 	i = 0;
 	while (i < philo->info->nop)
 	{
+		if (first_dead_philo == pids[i])
+		{
+			++i;
+			continue ;
+		}
 		if (!kill(pids[i], SIGTERM))
 			waitpid(-1, &status, 0);
 		++i;
 	}
+	i = philo->info->nop;
+	while (i--)
+		sem_post(philo->shared->full_philos);
 	(void)full_monitor;
-	pthread_join(*full_monitor, NULL);
 	return (0);
 }
 
@@ -46,7 +53,6 @@ int	run_simulation(t_philo *philo)
 	pids = (pid_t *)malloc(sizeof(pid_t) * philo->info->nop);
 	if (pids == NULL)
 		return (1);
-	philo->pids = pids;
 	philo->info->start_at = get_current_time_ms();
 	philo->last_eat_time = philo->info->start_at;
 	i = 0;
@@ -59,9 +65,8 @@ int	run_simulation(t_philo *philo)
 		else if (pids[i++] == 0)
 			exit(do_routine(philo));
 	}
-	if (philo->info->nome)
-		if (pthread_create(&full_monitor, NULL, monitor_full, philo) != 0)
+	if (monitor_full_philos(&full_monitor, philo, pids))
 			return (return_with_free(pids, 1));
-	wait_end(philo, pids, &full_monitor);
+	wait_end(philo, pids, full_monitor);
 	return (return_with_free(pids, 0));
 }
